Add init tool for qat checkpoint.
